+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+* CLI -n, --namehash flag
 
 ## 1.3.2
 * domains like "hello..crypto", "hello..eth", "hello..zil" should throw ResolutionErrorCode.UnsupportedDomain